fix modal response
