+++ SubsClient/Sources/Menu/Menu.swift
@@ -1,15 +1,15 @@
 //
-//  RecommendSubscriptionList.swift
+//  Menu.swift
 //  SubsClient
 //
-//  Created by 長田卓馬 on 2020/06/09.
+//  Created by 長田卓馬 on 2020/06/14.
 //
 
 import ComposableArchitecture
 import Core
 import Foundation
 
-struct RecommendSubscriptionList {
+struct Menu {
     static let reducer = Reducer<State, Action, Environment> { state, action, environment in
         switch action {
         case .fetchRecommendSubscriptions:
@@ -23,10 +23,23 @@ struct RecommendSubscriptionList {
                 .map(Action.recommendSubscriptionsResponse)
                 .cancellable(id: ID(), cancelInFlight: true)
         case let .recommendSubscriptionsResponse(.success(subscriptions)):
-            state.subscriptions = subscriptions
+            state.recommendSubscriptions = subscriptions
             return .none
         case let .recommendSubscriptionsResponse(.failure(error)):
-            state.subscriptions = []
+            state.recommendSubscriptions = []
+            return .none
+        case .fetchPopularSubscriptions:
+            return environment.popularSubscriptionsRepository
+                .fetch()
+                .receive(on: environment.mainQueue)
+                .catchToEffect()
+                .map(Action.popularSubscriptionsResponse)
+                .cancellable(id: ID(), cancelInFlight: true)
+        case let .popularSubscriptionsResponse(.success(subscriptions)):
+            state.popularSubscriptions = subscriptions
+            return .none
+        case let .popularSubscriptionsResponse(.failure(error)):
+            state.popularSubscriptions = []
             return .none
         }
     }.debug()
@@ -35,20 +48,24 @@ struct RecommendSubscriptionList {
 
     struct State: Equatable {
         static func == (lhs: State, rhs: State) -> Bool {
-            lhs.subscriptions == rhs.subscriptions
+            lhs.recommendSubscriptions == rhs.recommendSubscriptions
         }
 
-        var subscriptions: [Subscription_Subscription] = []
+        var recommendSubscriptions: [Subscription_Subscription] = []
+        var popularSubscriptions: [Subscription_Subscription] = []
     }
 
     enum Action {
         case fetchRecommendSubscriptions
         case recommendSubscriptionsResponse(Result<[Subscription_Subscription], Error>)
+        case fetchPopularSubscriptions
+        case popularSubscriptionsResponse(Result<[Subscription_Subscription], Error>)
     }
 
     struct Environment {
         let firebaseRepository: FirebaseRepository
         let recommendSubscriptionsRepository: AnyMySubscriptionServiceRequestable<[Subscription_Subscription]>
+        let popularSubscriptionsRepository: AnySubscriptionServiceRequestable<[Subscription_Subscription]>
         let mainQueue: AnySchedulerOf<DispatchQueue>
     }
 }
