+++ SubsClient/Sources/Menu/MenuView.swift
@@ -14,20 +14,12 @@ struct MenuView: View {
     @State private var showModal: Bool = false
     private let tabs = MenuTab.allCases
 
-    private let recommendStore = Store(
-        initialState: RecommendSubscriptionList.State(),
-        reducer: RecommendSubscriptionList.reducer,
-        environment: RecommendSubscriptionList.Environment(
+    private let store = Store<Menu.State, Menu.Action>(
+        initialState: Menu.State(),
+        reducer: Menu.reducer,
+        environment: Menu.Environment(
             firebaseRepository: AppEnvironment.shared.firebaseRepository,
             recommendSubscriptionsRepository: AppEnvironment.shared.recommendSubscriptionsRepository,
-            mainQueue: AppEnvironment.shared.mainQueue
-        )
-    )
-
-    private let popularStore = Store(
-        initialState: PopularSubscriptionList.State(),
-        reducer: PopularSubscriptionList.reducer,
-        environment: PopularSubscriptionList.Environment(
             popularSubscriptionsRepository: AppEnvironment.shared.popularSubscriptionsRepository,
             mainQueue: AppEnvironment.shared.mainQueue
         )
@@ -42,41 +34,47 @@ struct MenuView: View {
     }
 
     var body: some View {
-        NavigationView {
-            VStack(alignment: .leading) {
-                SlidingTabView(selection: $selectedTabIndex, tabs: tabs.map { $0.title })
-                if selectedTabIndex == 0 {
-                    RecommendSubscriptionListView(store: recommendStore)
-                } else if selectedTabIndex == 1 {
-                    PopularSubscriptionListView(store: popularStore)
+        WithViewStore(self.store) { viewStore in
+            NavigationView {
+                VStack(alignment: .leading) {
+                    SlidingTabView(selection: self.$selectedTabIndex, tabs: self.tabs.map { $0.title })
+                    if self.selectedTabIndex == 0 {
+                        RecommendSubscriptionListView(subscriptions: viewStore.recommendSubscriptions)
+                    } else if self.selectedTabIndex == 1 {
+                        PopularSubscriptionListView(subscriptions: viewStore.popularSubscriptions)
+                    }
                 }
-            }
-            .navigationBarTitle("選択する", displayMode: .inline)
-            .navigationBarItems(
-                trailing: Button(action: {
-                    self.showModal = true
-                }, label: {
-                    Image(systemName: "paperplane.fill")
-                        .foregroundColor(.primary)
-                        .font(.system(size: 20))
-                })
-                    .sheet(
-                        isPresented: self.$showModal,
-                        content: {
-                            SubscriptionCreateView(
-                                store: .init(
-                                    initialState: .init(),
-                                    reducer: SubscriptionCreate.reducer,
-                                    environment: SubscriptionCreate.Environment(
-                                        subscriptionRepository: AppEnvironment.shared.subscriptionRepository,
-                                        firebaseRepository: AppEnvironment.shared.firebaseRepository,
-                                        mainQueue: AppEnvironment.shared.mainQueue
+                .navigationBarTitle("選択する", displayMode: .inline)
+                .navigationBarItems(
+                    trailing: Button(action: {
+                        self.showModal = true
+                    }, label: {
+                        Image(systemName: "paperplane.fill")
+                            .foregroundColor(.primary)
+                            .font(.system(size: 20))
+                    })
+                        .sheet(
+                            isPresented: self.$showModal,
+                            content: {
+                                SubscriptionCreateView(
+                                    store: .init(
+                                        initialState: .init(),
+                                        reducer: SubscriptionCreate.reducer,
+                                        environment: SubscriptionCreate.Environment(
+                                            subscriptionRepository: AppEnvironment.shared.subscriptionRepository,
+                                            firebaseRepository: AppEnvironment.shared.firebaseRepository,
+                                            mainQueue: AppEnvironment.shared.mainQueue
+                                        )
                                     )
                                 )
-                            )
-                        }
-                    )
-            )
+                            }
+                        )
+                )
+            }
+            .onAppear {
+                viewStore.send(.fetchRecommendSubscriptions)
+                viewStore.send(.fetchPopularSubscriptions)
+            }
         }
     }
 }
